+++ CHANGES.md
@@ -9,6 +9,8 @@ v1.4.0 (unreleased)
 * Modify profile viewer so that when in 'Sum' mode, parts of profiles
   with no valid values are NaN rather than zero.
 
+* Add support for using degrees in full-sphere projections. [#2279]
+
 v1.3.0 (2022-04-22)
 -------------------